bugfix: ensure the room is not partially joined
Else with large fed latency workers may /leave and re-/join
before the partial join has completed. This trips up code in
`_should_perform_remote_join` which states:
```py
        # If we are not fully joined yet, and the target is not already in the room,
        # let's do a remote join so another server with the full state can validate
        # that the user has not been banned for example.
        # We could just accept the join and wait for state res to resolve that later on
        # but we would then leak room history to this person until then, which is pretty
        # bad.
        if is_partial_state_room and previous_membership != Membership.JOIN:
            return True, remote_room_hosts
```
This fails because workers do not supply server names
as it assumes the server is actually joined to the room via
the master (which is true).